Move grpc.scala to api package
