+++ node/src/main/scala/coop/rchain/node/node.scala
@@ -28,6 +28,7 @@ import diagnostics.MetricsServer
 import coop.rchain.comm.transport._
 import coop.rchain.comm.discovery._
 import coop.rchain.shared._
+import coop.rchain.node.api._
 
 import scala.io.Source
 import scala.util.{Failure, Success, Try}
